+++ src/core/mod.rs
@@ -7,5 +7,6 @@ mod clocks;
 mod duration_name;
 mod general_midi;
 
+pub use clocks::Clocks;
 pub use duration_name::DurationName;
 pub use general_midi::GeneralMidi;
